Remove const requirement for received data.
NOTE this ultimately stems from the use of Fw::Buffer and its related
function 'getSerializeRepr()' in the data field constructor.
Using Fw::Buffer / Fw::SerializeBufferBase opens up a lot of powerful
mechanisms for easily serializing data and passing it around without
copies but we should consider if that can be done with a const user input.